Move logging context map into `log.py` module
